Souls System
Added the souls system by creating a script for a souls manager and created a souls sphere object. Enemies now have values for the number of souls they carry which is defined in the spawn data within the resource foulder. Also made changes to the player to fix his jumping.