Player Slowdown, removed jumping
When the player gets hit, the animation speed and the player speed are reduced by a multiplier. A timer is set and when that timer goes off then the speed is set back to normal. The timer goes for as long as the blinking effect is on the player. Also removed jumping from the player.